requirements.txt
# FastAPI 핵심 패키지
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 데이터베이스 (MySQL)
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1

# 캐싱 (Redis)
redis==5.0.1
aioredis==2.0.1

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 환경 변수 관리
python-dotenv==1.0.0

# HTTP 클라이언트 (거래소 API)
httpx==0.25.2
aiohttp==3.9.1

# 데이터 검증
pydantic==2.5.0
pydantic-settings==2.1.0

# 유틸리티
python-dateutil==2.8.2
pytz==2023.3

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# 모니터링
prometheus-client==0.19.0

# WebSocket
websockets==12.0

# AWS 관련 패키지
boto3==1.34.0
botocore==1.34.0

# 모니터링 관련 패키지
psutil==5.9.6boto3==1.34.0
botocore==1.34.0
aiosqlite
